+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add option to set the alignment of the editor content when a max. width is set
+
 ### Fixed
 
 - Fix panel dividers in front of insert hyperlink modal
+- Theme doesn't take into account max-width introduced in latest joplin version
 
 ## [1.1.0] - 2021-08-16